Colocate actions so they're validated
